6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float f;
 
-	printf("Size of a char: %i byte(s)\n", sizeof(a));
-	printf("Size of an int: %i byte(s)\n", sizeof(b));
-	printf("Size of a long int: %i byte(s)\n", sizeof(c));
-	printf("Size of a long long int: %i byte(s)\n", sizeof(d));
-	print("Sise of a float: %i byte(s)\n", sizeof(f));
+	printf("Size of a char: %i byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %i byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a long int: %i byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %i byte(s)\n", (unsigned long)sizeof(d));
+	print("Sise of a float: %i byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
